Disable caching
Looks like the GitHub actions CI environment has been running out of space during the build recently and failing the job as a result. These are some options to not save downloaded packages and build stages to the disk in the hope it leaves us with more free space.

These options are useless in single use CI environments anyway as the cache is gone on the next build.