Remove lupin-support from pool
This doesn't exist in `jammy` anymore and is causing the daily builds to fail as a result. There's not a lot of information about it online but it's something to do with being able to install with the iso mounted as a loop filesystem. Ubuntu have dropped it, and I highly doubt it's compatible with our installer anyway :shrug: